+++ tybotRoute/test/conversation-intent_with_json_attributes_bot.js
@@ -1,0 +1,47 @@
+const bot = {
+	"webhook_enabled": false,
+	"language": "en",
+	"name": "Filter bot",
+	"type": "tilebot",
+	"intents": [{
+		"webhook_enabled": false,
+		"enabled": true,
+		"actions": [{
+			"_tdActionType": "reply",
+			"text": "Hi",
+			"attributes": {
+				"commands": [{
+					"type": "wait",
+					"time": 500
+				}, {
+					"type": "message",
+					"message": {
+						"type": "text",
+						"text": "Hi"
+					}
+				}]
+			}
+		}],
+		"question": "\\start",
+		"intent_display_name": "start",
+		"language": "en",
+		"intent_id": "72d059bc-b2cb-4629-9ce0-36f602a22479"
+	}]
+}
+
+// normalize the bot structure for the static intent search
+let intents = bot.intents;
+delete bot.intents;
+// console.log ("bot still is", JSON.stringify(bot));
+// console.log ("bintents still are", intents[0]);
+intent_dict = {};
+for (let i = 0; i < intents.length; i++) {
+  intent_dict[intents[i].intent_display_name] = intents[i];
+}
+bot.intents = intent_dict;
+const bots_data = {
+  "bots": {}
+}
+bots_data.bots["botID"] = bot;
+
+module.exports = { bots_data: bots_data };